// atlassian-frontend/packages/dragonfruit/search-dialog/src/confluence/clients/response-types.tsx
export enum Scope {
  ConfluencePageBlogAttachment = 'confluence.page,blogpost,attachment',
  ConfluencePageBlog = 'confluence.page,blogpost',
  ConfluenceSpace = 'confluence.space',
  People = 'cpus.user',
  UserConfluence = 'urs.user-confluence',
}
// ...
export const ConfluenceSpace: 'confluence-space' = 'confluence-space' as const;
// ...
interface UrsPersonResponseItem {
  absoluteUrl?: string; // For any TCS errors this field will not be present in the response.
  avatarUrl: string;
  entityType: string;
  id: string;
  name: string;
  nickname?: string;
}

export interface ConfUrsPeopleResponse
  extends ScopedAggregatorResponse<Scope.UserConfluence> {
  results: UrsPersonResponseItem[];
}

interface PersonResponseItem {
  absoluteUrl?: string; // For any TCS errors this field will not be present in the response.
  account_id: string;
  name: string;
  nickname?: string;
  job_title?: string;
  picture: string;
}

export interface ConfPeopleResponse
  extends ScopedAggregatorResponse<Scope.People> {
  results: PersonResponseItem[];
}
// ...
export type ConfluenceItemContentType = 'page' | 'blogpost' | 'space';

interface Container {
  title: string;
  id?: string; // This has to be optional because Confluence doesn't return it
  displayUrl?: string;
}

interface ConfItemResponseItem {
  title: string; // this is highlighted
  baseUrl: string;
  url: string;
  content?: {
    id: string;
    type: ConfluenceItemContentType;
    space?: {
      id: string;
    };
  };
  container: Container;
  space?: {
    key: string; // currently used as instance-unique ID
    icon: {
      path: string;
    };
  };
  iconCssClass: string; // icon-file-* for attachments, otherwise not needed
  lastModified: string; // e.g. "2019-07-08T02:54:38.822Z"
}

export interface ConfItemResponse
  extends ScopedAggregatorResponse<
    Scope.ConfluencePageBlogAttachment | Scope.ConfluencePageBlog
  > {
  results: ConfItemResponseItem[];
  size?: number;
}

export interface ConfItemResult extends Result {
  containerName: string;
  containerId: string;
  resultType: typeof ConfluenceObjectResult;
  iconClass?: string;
  lastModified: string | undefined;
  // used to indicate the result came from the recently viewed FE cache
  isRecentResult: boolean;
}

export interface ConfItemResults extends Results<ConfItemResult> {
  totalSize: number;
}

export interface ConfSpaceResponse
  extends ScopedAggregatorResponse<Scope.ConfluenceSpace> {
  results: ConfItemResponseItem[];
}